requirements.txt
asgiref==3.3.1
Django==3.1.4
djangorestframework==3.12.2
gunicorn==20.0.4
numpy==1.19.4
pandas==1.1.4
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
